requirements.txt
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
bootstrap4==0.1.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-heroku==0.3.1
django-on-heroku==1.1.2
djangorestframework==3.12.4
drf-yasg==1.20.0
eth-hash==0.3.2
eth-typing==2.3.0
executing==0.8.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
ipython==8.0.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
netaddr==0.8.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.27
psycopg2==2.9.2
psycopg2-binary==2.9.3
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.6
pytz==2021.3
pywin32==303
requests==2.26.0
ruamel.yaml==0.17.17
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
tomli==2.0.0
toolz==0.11.2
traitlets==5.1.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
varint==1.0.2
wcwidth==0.2.5
whitenoise==5.3.0
